# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "open-cam API",
    "description": "open-cam 是视频流分析与监控管理平台。\n\n- **摄像头**：接入局域网 RTSP 流或本地视频文件，YOLO（cuda/mps/cpu 自适应）检测与跟踪。\n- **规则引擎**：区域入侵 / 徘徊逗留 / 人数统计 / 区域人数 / 越线计数，支持生效时段与冷却去抖。\n- **事件**：命中规则即落库并存快照，可经 VLM（OpenAI 兼容接口）异步复核；处置闭环支持关注星标、负责人指派、状态流转与 webhook 通知，全程留痕。\n- **方案包**：行业规则模板包（连锁零售 / 美容美发 / 餐饮 / 快餐），一键安装与应用。\n- **统计**：分时段进出店客流等聚合视图。\n\n交互式调试见 [Swagger UI](/docs)，结构化阅读见 [ReDoc](/redoc)。\n",
    "contact": {
      "name": "open-cam",
      "url": "https://github.com/local/open-cam"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/training/tasks/{task_id}/train": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "启动本地微调训练（后台执行）",
        "description": "从预训练 YOLO 分类模型微调，产出 best.pt 与评估报告；评估集必含人工确认样本。状态用 GET 同路径查询。",
        "operationId": "start_train_api_training_tasks__task_id__train_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest",
                "default": {
                  "epochs": 20,
                  "imgsz": 224,
                  "val_ratio": 0.2
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "training"
        ],
        "summary": "查询训练状态与评估报告",
        "description": "训练完成后 result 即人话报告（含三指标与典型对错样本）。",
        "operationId": "train_status_api_training_tasks__task_id__train_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/assets": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "模型资产列表",
        "operationId": "list_assets_api_models_assets_get",
        "parameters": [
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "name": "model_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Kind"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按名称或描述搜索",
              "title": "Q"
            },
            "description": "按名称或描述搜索"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelAssetOut"
                  },
                  "title": "Response List Assets Api Models Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "创建模型资产",
        "description": "登记模型名称、描述、来源类型和模型能力；不等于部署到运行时。",
        "operationId": "create_asset_api_models_assets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelAssetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAssetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/assets/{asset_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "模型资产详情",
        "operationId": "get_asset_api_models_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAssetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "models"
        ],
        "summary": "更新模型资产",
        "operationId": "update_asset_api_models_assets__asset_id__patch",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelAssetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAssetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/assets/{asset_id}/bindings": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "查看模型关联",
        "operationId": "list_bindings_api_models_assets__asset_id__bindings_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelBindingOut"
                  },
                  "title": "Response List Bindings Api Models Assets  Asset Id  Bindings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "关联模型",
        "description": "支持手工关联，也支持保存 AI 推荐关联的置信度和理由；当前不自动改变运行时流水线。",
        "operationId": "create_binding_api_models_assets__asset_id__bindings_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelBindingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelBindingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/assets/{asset_id}/bindings/{binding_id}": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "解除模型关联",
        "operationId": "delete_binding_api_models_assets__asset_id__bindings__binding_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Asset Id"
            }
          },
          {
            "name": "binding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Binding Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "登记模型版本",
        "description": "记录指标、产物路径与来源任务；不改变线上模型。",
        "operationId": "register_api_models_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "models"
        ],
        "summary": "模型版本列表",
        "operationId": "list_models_api_models_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "name": "slot_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slot Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelVersionOut"
                  },
                  "title": "Response List Models Api Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "模型版本详情",
        "operationId": "get_model_api_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/deploy": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "部署模型",
        "description": "与线上模型做 A/B 对比，仅在三项指标全面更优时建议并执行替换；未更优返回 409，force=true 可强行上线。旧 live 降为 previous，回滚入口常驻。",
        "operationId": "deploy_api_models__model_id__deploy_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployBody",
                "default": {
                  "force": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/rollback": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "回滚到上一线上版本",
        "description": "按该版本所在槽位恢复 previous；没有上一版本时 400。",
        "operationId": "rollback_api_models__model_id__rollback_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/compare": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "与线上模型对比（不部署）",
        "operationId": "compare_api_models__model_id__compare_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeployBody": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "未全面更优时仍强制替换；默认拒绝以免训出更差模型上线",
            "default": false
          }
        },
        "type": "object",
        "title": "DeployBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelAssetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "source_type": {
            "type": "string",
            "pattern": "^(builtin|published|solution|uploaded|trained)$",
            "title": "Source Type"
          },
          "model_kind": {
            "type": "string",
            "pattern": "^(object_detection|classification|segmentation|pose|ocr|vlm)$",
            "title": "Model Kind"
          },
          "task_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Key"
          },
          "solution_pack_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution Pack Id"
          },
          "training_task_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Task Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "name",
          "source_type",
          "model_kind"
        ],
        "title": "ModelAssetCreate"
      },
      "ModelAssetOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "model_kind": {
            "type": "string",
            "title": "Model Kind"
          },
          "task_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Key"
          },
          "solution_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution Pack Id"
          },
          "training_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Task Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "source_type",
          "model_kind",
          "task_key",
          "solution_pack_id",
          "training_task_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ModelAssetOut"
      },
...
      "TrainRequest": {
        "properties": {
          "epochs": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Epochs",
            "description": "微调轮数",
            "default": 20
          },
          "imgsz": {
            "type": "integer",
            "maximum": 1280.0,
            "minimum": 32.0,
            "title": "Imgsz",
            "description": "训练输入边长",
            "default": 224
          },
          "val_ratio": {
            "type": "number",
            "exclusiveMaximum": 0.5,
            "exclusiveMinimum": 0.0,
            "title": "Val Ratio",
            "description": "自动标注样本进评估集的比例（确认样本必进评估集）",
            "default": 0.2
          }
        },
        "type": "object",
        "title": "TrainRequest"
      },
...
    }
  },
...
}